fix(createPledge): Use refetchQuery
